pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scrivai"
version = "0.1.0"
description = "可配置的通用文档生成与审核框架"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "chengzhang yu", email = "iomgaaycz@gmail.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Linguistic",
]
dependencies = [
    "litellm>=1.0",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "qmd>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.urls]
Homepage = "https://github.com/iomgaa/scrivai"
Repository = "https://github.com/iomgaa/scrivai"
Issues = "https://github.com/iomgaa/scrivai/issues"

[tool.setuptools.packages.find]
include = ["scrivai*", "utils*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
markers = [
    "integration: 真实 API 调用测试（需要网络和 API key）",
    "unit: 单元测试（使用 mock）",
]
addopts = "-v --tb=short"
